AnemicDetect_BE/requirements.txt
Django>=4.0
djangorestframework>=3.14
torch
torchvision
Pillow
ultralytics
opencv-python
numpy